requirements.txt
# Pinned dependencies that cause breakage
Werkzeug==3.0.1
SQLAlchemy==2.0.0

# Runtime dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.2
psycopg[binary]==3.1.12
retry2==0.9.5
python-dotenv==1.0.0

# Runtime tools
gunicorn==21.2.0
honcho==1.1.0

# Code quality
pylint==2.17.5
flake8==6.1.0
black==23.7.0

# Testing dependencies
pytest==7.4.0
pytest-pspec==0.0.4
pytest-cov==4.1.0
green==3.4.3
factory-boy==3.3.0
coverage==7.3.0

# Utilities
httpie==3.2.2

# BDD
behave==1.2.6
selenium==4.10.0
compare==0.2b0
requests==2.31.0